Switches the domain
