+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,5 +1,5 @@
 blank_issues_enabled: false
 contact_links:
   - name: awork Support
-    url: support@awork.io
+    url: support@awork.com
     about: Please ask none API related questions about awork here.
