+++ examples/python/using-oauth.py
@@ -1,9 +1,9 @@
 #!/usr/bin/env python3
 #
-# a simple example to query the API of awork.io with Python3 and OAuth2.0
+# a simple example to query the API of awork.com with Python3 and OAuth2.0
 # 
-# Website: https://www.awork.io/
-# Docs: https://developers.awork.io/
+# Website: https://www.awork.com/
+# Docs: https://developers.awork.com/
 # Forked from: https://requests-oauthlib.readthedocs.io/en/latest/examples/outlook.html
 #
 # Hint: pip3 install requests_oauthlib
@@ -16,11 +16,11 @@ from requests_oauthlib import OAuth2Session
 CLIENT_ID = '<your client id>'
 CLIENT_SECRET = '<your client secret>'
 
-# OAuth endpoints & params from API Description (https://developers.awork.io/)
-AUTHORIZATION_BASE_URL = 'https://api.awork.io/api/v1/accounts/authorize'
-TOKEN_URL = 'https://api.awork.io/api/v1/accounts/token'
+# OAuth endpoints & params from API Description (https://developers.awork.com/)
+AUTHORIZATION_BASE_URL = 'https://api.awork.com/api/v1/accounts/authorize'
+TOKEN_URL = 'https://api.awork.com/api/v1/accounts/token'
 SCOPE = ['offline_access']
-REDIRECT_URI = 'https://www.awork.io'
+REDIRECT_URI = 'https://www.awork.com'
 
 # Merge the OAuth session string
 awork = OAuth2Session(CLIENT_ID, scope=SCOPE, redirect_uri=REDIRECT_URI)
@@ -40,5 +40,5 @@ print(token)
 print()
 
 # Fetch a protected resource, i.e. all users
-response = awork.get('https://api.awork.io/api/v1/users')
+response = awork.get('https://api.awork.com/api/v1/users')
 print(response.content)
